Improved MCP server integration
