+++ extensions/cursor/src/mcp/mcpService.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
 import { captureException } from '../sentry';
+import { getOpikApiKey } from '../utils';
 
 export class MCPService {
     private context: vscode.ExtensionContext;
@@ -12,7 +13,7 @@ export class MCPService {
     /**
      * Register the Opik MCP server with Cursor
      */
-    async registerServer(apiKey: string): Promise<boolean> {
+    async registerServer(): Promise<boolean> {
         try {
             // Get configuration values
             const config = vscode.workspace.getConfiguration();
@@ -20,14 +21,15 @@ export class MCPService {
             const apiUrl = config.get<string>('opik.apiUrl', 'https://www.comet.com/opik/api');
             const workspace = config.get<string>('opik.workspace', 'default');
 
+            const apiKey = getOpikApiKey();
             if (!mcpEnabled) {
                 console.log('MCP server registration is disabled in settings');
                 return false;
             }
 
-            console.log('Registering Opik MCP server...');
-
             // Register the server using Cursor's MCP extension API
+            await (vscode as any).cursor.mcp.unregisterServer(this.serverName);
+
             await (vscode as any).cursor.mcp.registerServer({
                 name: this.serverName,
                 server: {
@@ -35,8 +37,7 @@ export class MCPService {
                     args: [
                         '-y',
                         'opik-mcp',
-                        '--apiKey',
-                        apiKey,
+                        ...(apiKey ? ['--apiKey', apiKey] : []),
                         '--apiUrl',
                         apiUrl,
                         '--workspace',
@@ -45,14 +46,6 @@ export class MCPService {
                     env: {}
                 }
             });
-
-
-            console.log('✅ Opik MCP server registered successfully');
-
-            // Show success message to user
-            vscode.window.showInformationMessage(
-                'Opik MCP server registered! You can now use Opik context in your chats.'
-            );
 
             return true;
 
@@ -69,35 +62,5 @@ export class MCPService {
         }
     }
 
-    /**
-     * Unregister the Opik MCP server
-     */
-    async unregisterServer(): Promise<boolean> {
-        try {
-            console.log('Unregistering Opik MCP server...');
-
-            await (vscode as any).cursor.mcp.unregisterServer(this.serverName);
-            console.log('✅ Opik MCP server unregistered successfully');
-            return true;
-
-        } catch (error) {
-            captureException(error);
-            console.error('❌ Failed to unregister Opik MCP server:', error);
-            return false;
-        }
-    }
-
-    /**
-     * Re-register the server with a new API key
-     */
-    async reregisterServer(apiKey: string): Promise<boolean> {
-        console.log('Re-registering Opik MCP server with new API key...');
-
-        // First unregister the existing server
-        await this.unregisterServer();
-
-        // Then register with new API key
-        return await this.registerServer(apiKey);
-    }
 
 }
